requirements.txt
huggingface-hub==0.17.1
llama-cpp-python==0.1.78
streamlit==1.26.0
pyngrok==4.1.1